♻️ Made get_current_optimization_iteration a static method of Optimizer
